:sparkles: Added processing forever

+++ src/dataset/download_models.py
@@ -4,7 +4,7 @@ import time
 import webbrowser
 
 from src.config import THINGIVERSE_FLASK_PORT, THINGIVERSE_FLASK_WAIT_PRE, THINGIVERSE_FLASK_ENDPOINT, \
-    THINGIVERSE_FLASK_WAIT_POST
+    THINGIVERSE_FLASK_WAIT_POST, THINGIVERSE_FLASK_WAIT_ENABLE
 from src.helper import log
 from src.thingiverse import launcher
 
@@ -20,11 +20,14 @@ def main():
     log.info(f'Sleeping {THINGIVERSE_FLASK_WAIT_PRE} seconds for letting the API to initialize and open a browser.')
     time.sleep(THINGIVERSE_FLASK_WAIT_PRE)
     webbrowser.open(f'http://localhost:{THINGIVERSE_FLASK_PORT}{THINGIVERSE_FLASK_ENDPOINT}', new=2)
-    log.info(f'Sleeping {THINGIVERSE_FLASK_WAIT_POST} seconds for letting the process to finish.')
-    log.info('You can stop the script if the process is done.')
-    time.sleep(THINGIVERSE_FLASK_WAIT_POST)
-    log.info('Killing processes.')
-    process.terminate()
+    if THINGIVERSE_FLASK_WAIT_ENABLE:
+        log.info(f'Sleeping {THINGIVERSE_FLASK_WAIT_POST} seconds for letting the process to finish.')
+        log.info('You can stop the script if the process is done.')
+        time.sleep(THINGIVERSE_FLASK_WAIT_POST)
+        log.info('Killing processes.')
+        process.terminate()
+    else:
+        log.info('Processing forever.')
 
 
 if __name__ == '__main__':
